Harrison/multi vector (#9700)
