+++ libs/langchain/langchain/retrievers/__init__.py
@@ -40,6 +40,7 @@ from langchain.retrievers.merger_retriever import MergerRetriever
 from langchain.retrievers.metal import MetalRetriever
 from langchain.retrievers.milvus import MilvusRetriever
 from langchain.retrievers.multi_query import MultiQueryRetriever
+from langchain.retrievers.multi_vector import MultiVectorRetriever
 from langchain.retrievers.parent_document_retriever import ParentDocumentRetriever
 from langchain.retrievers.pinecone_hybrid_search import PineconeHybridSearchRetriever
 from langchain.retrievers.pubmed import PubMedRetriever
@@ -92,4 +93,5 @@ __all__ = [
     "WebResearchRetriever",
     "EnsembleRetriever",
     "ParentDocumentRetriever",
+    "MultiVectorRetriever",
 ]
